requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
cloudinary==1.44.1
dj-database-url==3.1.2
Django==5.2.13
django-cloudinary-storage==0.3.0
djangorestframework==3.17.1
dlib==20.0.1
drf-spectacular==0.29.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==25.3.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.11
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
